+++ src/utils/getGeoCoordinates.js
@@ -2,6 +2,7 @@ import axios from "axios";
 
 const getGeoCoordinates = async (address) => {
   /* Given address (only first line is necessary), get long and lat of address*/
+  console.log(address);
   try {
     const { data } = await axios.get(
       `http://localhost:3001/geocode/?address=${address}`
